+++ Functions/News/NewsFunctions.cs
@@ -121,7 +121,7 @@ namespace HeroServer
             long id = -1;
             using (TransactionScope scope = new TransactionScope(TransactionScopeAsyncFlowOption.Enabled))
             {
-                registerNewsRequest.Post.PostTypeId = (long)PostType.News;
+                registerNewsRequest.Post.PostTypeId = PostType.News;
                 registerNewsRequest.News.PostId = await PostFunctions.Register(registerNewsRequest);
 
                 registerNewsRequest.News.Status = 1;
