+++ Functions/Product/ProductFunctions.cs
@@ -121,7 +121,7 @@ namespace HeroServer
             long id = -1;
             using (TransactionScope scope = new TransactionScope(TransactionScopeAsyncFlowOption.Enabled))
             {
-                registerProductRequest.Post.PostTypeId = (long)PostType.Product;
+                registerProductRequest.Post.PostTypeId = PostType.Product;
                 registerProductRequest.Product.PostId = await PostFunctions.Register(registerProductRequest);
 
                 registerProductRequest.Product.Status = 1;
